update store structure with redux-immutable
