+++ client/controllers/actions/message.js
@@ -32,19 +32,3 @@ export const messageActions = {
     },
   }),
 };
-
-export const [POST, postActions] = asyncActionBundle('@event/POST');
-
-export const EVENT = {
-  MESSAGE,
-  POST,
-  ASYNC_TEST,
-  TEST,
-};
-
-export const eventActions = {
-  messageActions,
-  postActions,
-  asyncTestActions,
-  testAction,
-};
