+++ client/controllers/actions/post.js
@@ -1,0 +1,8 @@
+import { asyncActionBundle } from 'utils/action-manager';
+
+export const [POST, postActions] = asyncActionBundle('@event/POST');
+
+export default {
+  POST,
+  postActions,
+};
