+++ client/controllers/logics/message.js
@@ -1,0 +1,18 @@
+import fetch from 'isomorphic-fetch';
+import { createLogic } from 'redux-logic';
+
+import { MESSAGE, messageActions } from 'controllers/actions/message';
+
+const getMessageLogic = createLogic({
+  type: MESSAGE.GET,
+  process({ getState }, dispatch, done) {
+    const index = getState().getIn(['count', 'current']);
+    fetch(`/message/${index}`)
+      .then(response => response.json())
+      .then(data => dispatch(messageActions.receive(data)))
+      .catch(err => dispatch(messageActions.error(err)))
+      .then(() => done());
+  },
+});
+
+export default [getMessageLogic];
