+++ client/controllers/reducers/index.js
@@ -1,13 +1,15 @@
-import { combineReducers } from 'redux';
+import { combineReducers } from 'redux-immutable';
 import { routerReducer as router } from 'react-router-redux';
+import { immutableAsyncReducers } from 'utils/action-manager';
+
+import { MESSAGE } from 'controllers/actions/message';
+import { POST } from 'controllers/actions/post';
 
 import count from './count';
-import event from './event';
 
-const reducer = combineReducers({
+export default combineReducers({
   router,
   count,
-  event,
+  message: immutableAsyncReducers(MESSAGE),
+  post: immutableAsyncReducers(POST),
 });
-
-export default reducer;
